bigger min-width, fake extra image width
The minimum width of the entire page has been changed from 400px
to 480px.
As the image have less width as I espected, the background from
it's parent was visible at the right, acting like a border, so
I added a linear gradient as background to the parent to fake
the image's continuation, same for it's border.